requirements.txt
alembic==1.9.1
anyio==3.6.2
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
ecdsa==0.18.0
fastapi==0.88.0
fastapi-jwt-auth==0.5.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
PyJWT==1.7.1
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a31
sqlmodel==0.0.8
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
